Moved update stalker logics from binder file.
