+++ src/engine/core/utils/logics.test.ts
@@ -1,0 +1,5 @@
+import { describe, it } from "@jest/globals";
+
+describe("updateStalkerLogic util", () => {
+  it.todo("should correctly update stalker logics");
+});
